organized codebase
