+++ data.json
@@ -7,7 +7,7 @@
           {
             "colType": "dated",
             "active": true,
-			"collapsed": false,
+            "collapsed": false,
             "data": {
               "name": "Over Due",
               "index": 1,
@@ -23,7 +23,7 @@
           {
             "colType": "dated",
             "active": true,
-			"collapsed": false,
+            "collapsed": false,
             "data": {
               "name": "Today",
               "index": 2,
@@ -39,7 +39,7 @@
           {
             "colType": "dated",
             "active": true,
-			"collapsed": false,
+            "collapsed": false,
             "data": {
               "name": "Tomorrow",
               "index": 3,
@@ -55,7 +55,7 @@
           {
             "colType": "dated",
             "active": true,
-			"collapsed": false,
+            "collapsed": false,
             "data": {
               "name": "Future",
               "index": 4,
@@ -71,7 +71,7 @@
           {
             "colType": "undated",
             "active": true,
-			"collapsed": false,
+            "collapsed": false,
             "data": {
               "name": "Undated",
               "index": 5
@@ -80,7 +80,7 @@
           {
             "colType": "completed",
             "active": true,
-			"collapsed": false,
+            "collapsed": false,
             "data": {
               "limit": 10,
               "name": "Completed",
@@ -104,7 +104,7 @@
           {
             "colType": "untagged",
             "active": true,
-			"collapsed": false,
+            "collapsed": false,
             "data": {
               "name": "Untagged",
               "index": 1
@@ -113,7 +113,7 @@
           {
             "colType": "namedTag",
             "active": true,
-			"collapsed": false,
+            "collapsed": false,
             "data": {
               "name": "Testing Tasks",
               "index": 2,
@@ -123,7 +123,7 @@
           {
             "colType": "namedTag",
             "active": true,
-			"collapsed": false,
+            "collapsed": false,
             "data": {
               "name": "GPT",
               "index": 3,
@@ -133,7 +133,7 @@
           {
             "colType": "namedTag",
             "active": true,
-			"collapsed": false,
+            "collapsed": false,
             "data": {
               "name": "Test Task",
               "index": 4,
@@ -143,7 +143,7 @@
           {
             "colType": "namedTag",
             "active": true,
-			"collapsed": false,
+            "collapsed": false,
             "data": {
               "name": "Note Till Task is Done",
               "index": 5,
@@ -153,7 +153,7 @@
           {
             "colType": "otherTags",
             "active": true,
-			"collapsed": false,
+            "collapsed": false,
             "data": {
               "name": "Other Tags",
               "index": 6
@@ -162,7 +162,7 @@
           {
             "colType": "completed",
             "active": true,
-			"collapsed": false,
+            "collapsed": false,
             "data": {
               "index": 7,
               "limit": 10,
@@ -172,7 +172,7 @@
           {
             "colType": "namedTag",
             "active": true,
-			"collapsed": false,
+            "collapsed": false,
             "data": {
               "name": "Test Tasks",
               "index": 8,
@@ -195,7 +195,7 @@
           {
             "colType": "dated",
             "active": true,
-			"collapsed": false,
+            "collapsed": false,
             "data": {
               "name": "Past Tasks",
               "index": 1,
@@ -210,7 +210,7 @@
           {
             "colType": "namedTag",
             "active": true,
-			"collapsed": false,
+            "collapsed": false,
             "data": {
               "name": "Test Name Tag Column",
               "index": 2,
@@ -220,7 +220,7 @@
           {
             "colType": "untagged",
             "active": true,
-			"collapsed": false,
+            "collapsed": false,
             "data": {
               "name": "Untagged Tasks",
               "index": 3
@@ -229,7 +229,7 @@
           {
             "colType": "otherTags",
             "active": true,
-			"collapsed": false,
+            "collapsed": false,
             "data": {
               "name": "Other Tags Tasks",
               "index": 4
@@ -254,12 +254,12 @@
       "filters": [
         "MyReadme.md, Daily/, New.md"
       ],
-      "firstDayOfWeek": "2",
+      "firstDayOfWeek": "1",
       "ignoreFileNameDates": false,
       "taskCompletionFormat": "Obsidian",
       "taskCompletionInLocalTime": true,
       "taskCompletionShowUtcOffset": true,
-      "autoAddDue": true,
+      "autoAddDue": false,
       "scanVaultAtStartup": false,
       "dayPlannerPlugin": true,
       "realTimeScanning": false
