+++ src/interfaces/GlobalVariables.ts
