+++ src/modal/BoardConfigModal.tsx
@@ -1,4 +1,4 @@
-// /src/compoenents/BoardConfigureModal.tsx - V2
+// /src/modal/BoardConfigModal.tsx - V2
 
 import { App, Modal, Notice } from "obsidian";
 import { DragDropContext, Draggable, Droppable } from "react-beautiful-dnd"; // For drag-and-drop
