+++ src/views/KanbanView.ts
@@ -1,12 +1,7 @@
 // src/views/KanbanView.ts
 
 import { App, ItemView, Vault, WorkspaceLeaf } from "obsidian";
-import {
-	loadBoardsData,
-	openBoardConfigModal,
-	openReScanVaultModal,
-	saveBoardsData,
-} from "../services/OpenColumnConfig";
+import { loadBoardsData, saveBoardsData } from "src/utils/SettingsOperations";
 
 import { Board } from "src/interfaces/KanbanBoard";
 import KanbanBoard from "../components/KanbanBoard";
@@ -14,7 +9,8 @@ import { ReScanVaultModal } from "src/modal/ReScanVaultModal";
 import React from "react";
 import ReactDOM from "react-dom/client";
 import TaskBoard from "../../main";
-import { VIEW_TYPE_TASKBOARD } from "src/interfaces/TaskBoardGlobalValues";
+import { VIEW_TYPE_TASKBOARD } from "src/interfaces/GlobalVariables";
+import { openReScanVaultModal } from "../services/OpenModals";
 
 export class KanbanView extends ItemView {
 	private vault: Vault;
@@ -52,7 +48,10 @@ export class KanbanView extends ItemView {
 			openReScanVaultModal(this.app);
 		});
 
-		console.log("The Settings which i have loaded using Obsidian : ", this.getSettings());
+		console.log(
+			"The Settings which i have loaded using Obsidian : ",
+			this.getSettings()
+		);
 
 		const root = ReactDOM.createRoot(this.contentEl); // Correct element reference
 		root.render(<KanbanBoard app={this.app} />); // Use 'this.app' here
@@ -76,14 +75,6 @@ export class KanbanView extends ItemView {
 		// Clean up when view is closed
 	}
 }
-
-
-
-
-
-
-
-
 
 // // src/views/KanbanView.ts   ----- Wokring - V2
